pyproject.toml
[project]
name = "assignment1-5"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0",
    "datasets>=2.14",
    "python-dotenv>=1.0",
    "loguru>=0.7",
]
